+++ example/lib/main.dart
@@ -1,42 +1,24 @@
 import 'package:flutter/material.dart';
 import 'package:link_rich_text/link_rich_text.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
+      title: 'link_rich_text demo',
       theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
         primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+      home: MyHomePage(title: 'link_rich_text demo'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
   MyHomePage({Key? key, this.title}) : super(key: key);
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
 
   final String? title;
 
@@ -45,115 +27,132 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
-    });
-  }
+  LinkRichModel? _linkRichModel;
 
   @override
-  Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
+  void initState() {
+    super.initState();
+    _setLinkRichModel();
+  }
+
+  _setLinkRichModel() async {
+    // 模拟加载数据的耗时
+    await Future.delayed(Duration(milliseconds: 500));
+
+    // 对于性能要求高的场景，比如列表中，
+    // 可在加载好数据的同时就设置好富文本模型（LinkRichModel）
+    // 并将模型传入LinkRichText中，或将模型的textSpan属性传入RichText中显示
+
+    // 添加特殊字符串
     List<SpecialStr> specialStrs = <SpecialStr>[];
     specialStrs.add(SpecialStr(
-        text: '老@老陈啊',
-        type: 'user',
-        style: TextStyle(fontSize: 15, color: Colors.purple)));
-    specialStrs.add(SpecialStr(
-        text: '@老陈',
+        text: '@我啊',
         type: 'user',
         style: TextStyle(fontSize: 15, color: Colors.blue)));
     specialStrs.add(SpecialStr(
-        text: '@老陈',
+        text: '@我啊啊',
         type: 'user',
-        style: TextStyle(fontSize: 15, color: Colors.orangeAccent)));
-    specialStrs.add(SpecialStr(
-        text: '@老陈啊',
-        type: 'user',
-        style: TextStyle(fontSize: 15, color: Colors.yellow)));
-    specialStrs.add(SpecialStr(
-        text: '@老陈啊啊',
-        type: 'user',
-        style: TextStyle(fontSize: 15, color: Colors.teal)));
+        style: TextStyle(fontSize: 18, color: Colors.purple)));
     specialStrs.add(SpecialStr(
         text: '@一只鱼',
         type: 'user',
-        style: TextStyle(fontSize: 15, color: Colors.blue)));
+        style: TextStyle(fontSize: 15, color: Colors.orange)));
     specialStrs.add(SpecialStr(
         text: '@不老实的鸟',
         type: 'user',
-        style: TextStyle(fontSize: 15, color: Colors.blue)));
+        style: TextStyle(fontSize: 15, color: Colors.brown)));
+
+    // 添加正则特殊字符串
     List<RegExpStr> regExpStrs = <RegExpStr>[];
     regExpStrs.add(RegExpStr(
         text: '#\\S*? ',
         type: '#',
-        style: TextStyle(fontSize: 15, color: Colors.lightBlueAccent)));
-    String text =
-        '@一只鱼 https://www.baidu.com 控件打底可反#馈的 了老@老陈啊都放到裤积分 @老陈 快进快手动，肯德的框架反馈的减@老陈啊啊肥看的积分基疯狂的发，发拉拉速度快积分@哈哈卢萨卡的积分禄口街道和你是谁 @不老实的鸟 ？是吗@老陈';
-//    text = '';
+        style: TextStyle(fontSize: 18, color: Colors.lightBlueAccent)));
+
+    String text = '@一只鱼 控件打底可反#馈的 了老@我啊都放到裤积分 '
+        '快进快手动，肯德的框架反馈的减@我啊啊肥看的积分基疯狂的发，'
+        '发拉拉速度快积分@哈哈卢萨卡的积分禄口街道和你是谁 @不老实的鸟 ？是吗';
+
+    // 创建并缓存富文本模型
+    _linkRichModel = LinkRichModel(
+      text,
+      style: TextStyle(fontSize: 15, color: Colors.black),
+      linkStyle: TextStyle(fontSize: 18, color: Colors.red),
+      specialStrs: specialStrs,
+      regExpStrs: regExpStrs,
+      onTapSpecialStr: (String text, String type) {
+        print('type = $type, text = $text');
+      },
+    );
+
+    setState(() {});
+  }
+
+  @override
+  Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
         title: Text(widget.title ?? ''),
       ),
       body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headline4,
-            ),
-            LinkRichText(
-              text,
-              style: TextStyle(fontSize: 15, color: Colors.black),
-              linkStyle: TextStyle(fontSize: 18, color: Colors.red),
-              specialStrs: specialStrs,
-              regExpStrs: regExpStrs,
-              onTapSpecialStr: (String text, String type) {
-                print('type = $type, text = $text');
-              },
-            ),
-          ],
+        child: _linkRichModel == null
+            ? Container()
+            : Column(
+                children: [
+                  _buildLinkRichText(),
+                  _buildSpan(),
+                  LinkRichText.fromModel(_linkRichModel!),
+                  _buildSpan(),
+                  RichText(text: _linkRichModel!.textSpan),
+                ],
+              ),
+      ),
+    );
+  }
+
+  _buildLinkRichText() {
+    // 添加特殊字符串
+    List<SpecialStr> specialStrs = <SpecialStr>[];
+    specialStrs.add(SpecialStr(
+        text: '@老大',
+        type: 'user',
+        style: TextStyle(fontSize: 15, color: Colors.purple)));
+
+    // 添加正则特殊字符串
+    List<RegExpStr> regExpStrs = <RegExpStr>[];
+    regExpStrs.add(RegExpStr(
+        text: '#\\S*? ',
+        type: '话题',
+        style: TextStyle(fontSize: 15, color: Colors.blueAccent)));
+
+    String text =
+        '@老大看看这个：https://pub.dev/packages/photo_browser，一个不错的#flutter 图片浏览器';
+
+    return LinkRichText(
+      text,
+      style: TextStyle(fontSize: 15, color: Colors.black),
+      linkStyle: TextStyle(fontSize: 15, color: Colors.blue),
+      specialStrs: specialStrs,
+      regExpStrs: regExpStrs,
+      onTapSpecialStr: (String text, String type) {
+        print('type = $type, text = $text');
+        if (type == 'link') {
+          launch(text);
+        }
+      },
+    );
+  }
+
+  _buildSpan() {
+    return SizedBox(
+      height: 20,
+      child: Center(
+        child: Container(
+          width: double.infinity,
+          height: 1,
+          color: Colors.blueGrey,
         ),
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }